--- Learning_test_automation/Spotkanie6_Zadania/BMI_vBeta/BMI.sln
@@ -1,43 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.5.33414.496
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Interface", "BMI\Interface.csproj", "{0AEBE9FA-BFC7-4DA1-A7CD-83C424C867B1}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Data", "Data\Data.csproj", "{2CA0AABA-1CFB-415B-BC9E-75D0CF0B761B}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Option", "Option\Option.csproj", "{C00B6AE6-7EA3-489D-847B-666F56113966}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BMI", "ClassLibrary1\BMI.csproj", "{B94DDA87-A59C-45D1-B760-5B48BA23FC89}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{0AEBE9FA-BFC7-4DA1-A7CD-83C424C867B1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{0AEBE9FA-BFC7-4DA1-A7CD-83C424C867B1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{0AEBE9FA-BFC7-4DA1-A7CD-83C424C867B1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{0AEBE9FA-BFC7-4DA1-A7CD-83C424C867B1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{2CA0AABA-1CFB-415B-BC9E-75D0CF0B761B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{2CA0AABA-1CFB-415B-BC9E-75D0CF0B761B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{2CA0AABA-1CFB-415B-BC9E-75D0CF0B761B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{2CA0AABA-1CFB-415B-BC9E-75D0CF0B761B}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C00B6AE6-7EA3-489D-847B-666F56113966}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C00B6AE6-7EA3-489D-847B-666F56113966}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C00B6AE6-7EA3-489D-847B-666F56113966}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C00B6AE6-7EA3-489D-847B-666F56113966}.Release|Any CPU.Build.0 = Release|Any CPU
-		{B94DDA87-A59C-45D1-B760-5B48BA23FC89}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{B94DDA87-A59C-45D1-B760-5B48BA23FC89}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{B94DDA87-A59C-45D1-B760-5B48BA23FC89}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{B94DDA87-A59C-45D1-B760-5B48BA23FC89}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {121A823E-F652-4505-A5EC-4014EDF7E780}
-	EndGlobalSection
-EndGlobal
