+++ Learning_test_automation/Spotkanie6_Zadania/BMI_vBeta/BMI/Announcement.cs
