+++ Learning_test_automation/Spotkanie6_Zadania/BMI_vBeta/BMI_vBeta/BMI_vBeta.csproj
@@ -1,9 +1,14 @@
-<Project Sdk="Microsoft.NET.Sdk">
+﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
+    <OutputType>Exe</OutputType>
     <TargetFramework>net6.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\BMI\BMI.csproj" />
+  </ItemGroup>
 
 </Project>
