+++ Learning_test_automation/Spotkanie6_Zadania/BMI_vBeta/BMI_vBeta/Program.cs
@@ -1,0 +1,16 @@
+﻿using BMI;
+
+namespace Interface
+{
+    class Program
+    {
+        static public void Main(string[] args)
+        {
+            while (true)
+            {
+                Announcement announcementVerdict = new Announcement();
+                announcementVerdict.ApplicationRunner();
+            }
+        }
+    }
+}
