--- Learning_test_automation/Spotkanie6_Zadania/BMI_vBeta/Interface/Program.cs
@@ -1,17 +1,0 @@
-﻿using BMI;
-
-namespace Interface
-{
-    class Program
-        {
-            static public void Main(string[] args)
-            {   
-                while (true)
-                {
-                    Announcement announcementVerdict = new Announcement();
-                    announcementVerdict.ApplicationRunner();
-                }
-            }
-        }
-}
-
