+++ Learning_test_automation/Spotkanie6_Zadania/BMI_vBeta/Option/Performer.cs
