--- Learning_test_automation/Spotkanie6_Zadania/BMI_vBeta/README.md
@@ -1,11 +1,0 @@
-Zadanie 2 - Aplikacja konsolowa do przeliczania BMI:
-Stwórz nową solucję oraz projekty, które uważasz, że będą potrzebne
-Wymagania aplikacji: 
-użytkownik podaje wagę
-użytkownik podaje wzrost
-skorzystaj ze wzoru do obliczania BMI
-
-BMI = waga / (wzrost)^2
-
-przyjmij zakres : 18,5 – 24,9  jako prawidłowy, a zakres poniżej jako niedowagę, a zakres powyżej jako nadwagę
-wyświetl informacje dla użytkownika
